requirements.txt
streamlit>=1.32.0
pandas>=2.2.3
numpy>=1.26.4
scikit-learn>=1.5.0
python-dotenv>=1.0.0
fastapi>=0.109.2
uvicorn>=0.27.1
python-multipart>=0.0.9
pymongo>=4.6.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.6.1
